Move _components directory from src/app/ to src/
