+++ src/app/episodes/[guid]/page.tsx
@@ -6,7 +6,7 @@ import { Metadata, ResolvingMetadata } from 'next'
 import '../../layout.css'
 import styles from './page.module.css'
 import { FeedLoader } from '../../_utils/FeedLoader';
-import { Episode } from '../../_components/types/Episode';
+import { Episode } from '../../../components/types/Episode';
 import { PublishedDate } from '../../_utils/PublishedDate';
 
 import sanitizeHtml from 'sanitize-html';
