+++ src/app/episodes/page/[page]/page.tsx
@@ -1,6 +1,6 @@
 import '../../../layout.css'
-import EpisodeEntry from '../../../_components/EpisodeEntry';
-import { Episode } from '../../../_components/types/Episode';
+import EpisodeEntry from '../../../../components/EpisodeEntry';
+import { Episode } from '../../../../components/types/Episode';
 import { FeedLoader } from '../../../_utils/FeedLoader';
 import Link from 'next/link';
 import React from 'react';
