+++ src/app/sitemap.ts
@@ -1,6 +1,6 @@
 import { MetadataRoute } from "next"
 import { FeedLoader } from './_utils/FeedLoader';
-import { Episode } from "./_components/types/Episode";
+import { Episode } from "../components/types/Episode";
 
 const BASE_URL = process.env.BASE_URL || '';
